refactor linodes.go to kernels, types, and instances
